=== Step 3 + Change Architecture + clean Code ===
