+++ app/src/App/Command/RegisterFleetCommand.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\App\Command;
+
+final class RegisterFleetCommand
+{
+    /**
+     * @param int $userId
+     */
+    public function __construct(
+        public int $userId
+    )
+    {
+    }
+}
