+++ app/src/App/Command/RegisterVehicleCommand.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\App\Command;
+
+final class RegisterVehicleCommand
+{
+    /**
+     * @param string $fleetId
+     * @param string $plateNumber
+     */
+    public function __construct(
+        public readonly string $fleetId,
+        public readonly string $plateNumber
+    )
+    {
+    }
+}
