--- app/src/Command/RegisterFleetCommand.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace App\Command;
-
-final class RegisterFleetCommand
-{
-    /**
-     * @param int $userId
-     */
-    public function __construct(
-        public int $userId
-    )
-    {
-    }
-}
