--- app/src/Controller/HomeController.php
@@ -1,15 +1,0 @@
-<?php
-namespace App\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
-
-final class HomeController extends AbstractController
-{
-    #[Route('/', name: 'homepage')]
-    public function index(): Response
-    {
-        return $this->render("home/index.html.twig");
-    }
-}
