+++ app/src/Domain/Interfaces/AlgoManagerInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Domain\Interfaces;
+
+interface AlgoManagerInterface
+{
+    /**
+     * @param int $index
+     * @return bool
+     */
+    public function algo(int $index = 100): bool;
+}
