+++ app/src/Domain/Interfaces/Entity/LocationInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Domain\Interfaces\Entity;
+
+/**
+ *
+ */
+interface LocationInterface
+{
+    /**
+     * @param LocationInterface $location
+     * @return bool
+     */
+    public function equals(LocationInterface $location): bool;
+}
