+++ app/src/Domain/Interfaces/Handler/ParkVehicleHandleInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Domain\Interfaces\Handler;
+
+interface ParkVehicleHandleInterface
+{
+
+}
