+++ app/src/Domain/Interfaces/Handler/RegisterVehicleHandlerInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Domain\Interfaces\Handler;
+
+interface RegisterVehicleHandlerInterface
+{
+
+}
