--- app/src/Interfaces/Entity/LocationInterface.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace App\Interfaces\Entity;
-
-/**
- *
- */
-interface LocationInterface
-{
-    /**
-     * @param LocationInterface $location
-     * @return bool
-     */
-    public function equals(LocationInterface $location): bool;
-}
