--- app/src/Interfaces/Entity/VehicleInterface.php
@@ -1,30 +1,0 @@
-<?php
-
-namespace App\Interfaces\Entity;
-
-/**
- *
- */
-interface VehicleInterface
-{
-    /**
-     * @return string
-     */
-    public function getPlateNumber(): string;
-
-    /**
-     * @return LocationInterface|null
-     */
-    public function getLocation(): ?LocationInterface;
-
-    /**
-     * @param LocationInterface $location
-     * @return void
-     */
-    public function parkAt(LocationInterface $location): void;
-
-    /**
-     * @return bool
-     */
-    public function isAlreadyOnFleet(): bool;
-}
