--- app/src/Interfaces/Handler/ParkVehicleHandleInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Interfaces\Handler;
-
-interface ParkVehicleHandleInterface
-{
-
-}
