--- app/src/Interfaces/Handler/RegisterFleetHandlerInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Interfaces\Handler;
-
-interface RegisterFleetHandlerInterface
-{
-
-}
