--- app/src/Interfaces/Handler/RegisterVehicleHandlerInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Interfaces\Handler;
-
-interface RegisterVehicleHandlerInterface
-{
-
-}
