--- app/src/Interfaces/Repository/VehicleRepositoryInterface.php
@@ -1,22 +1,0 @@
-<?php
-
-namespace App\Interfaces\Repository;
-
-use App\Interfaces\Entity\VehicleInterface;
-
-interface VehicleRepositoryInterface
-{
-
-    /**
-     * @param VehicleInterface $vehicle
-     * @return void
-     */
-    public function save(VehicleInterface $vehicle): void;
-
-    /**
-     * @param string $plateNumber
-     * @param int $fleetId
-     * @return VehicleInterface|null
-     */
-    public function findVehicle(string $plateNumber, int $fleetId): ?VehicleInterface;
-}
